Move introspection related classes out of web dto package into specific introspection package
Makes the auto configuration clearer and the purpose of these classes clearer.
