+++ genie-web/src/main/java/com/netflix/genie/web/agent/launchers/impl/LocalAgentLauncherImpl.java
@@ -20,10 +20,10 @@ package com.netflix.genie.web.agent.launchers.impl;
 import com.google.common.annotations.VisibleForTesting;
 import com.netflix.genie.web.agent.launchers.AgentLauncher;
 import com.netflix.genie.web.data.services.JobSearchService;
-import com.netflix.genie.web.dtos.GenieWebHostInfo;
-import com.netflix.genie.web.dtos.GenieWebRpcInfo;
 import com.netflix.genie.web.dtos.ResolvedJob;
 import com.netflix.genie.web.exceptions.checked.AgentLaunchException;
+import com.netflix.genie.web.introspection.GenieWebHostInfo;
+import com.netflix.genie.web.introspection.GenieWebRpcInfo;
 import com.netflix.genie.web.properties.LocalAgentLauncherProperties;
 import com.netflix.genie.web.util.ExecutorFactory;
 import io.micrometer.core.instrument.MeterRegistry;
