+++ genie-web/src/main/java/com/netflix/genie/web/introspection/GenieWebRpcInfo.java
@@ -15,14 +15,28 @@
  *     limitations under the License.
  *
  */
+package com.netflix.genie.web.introspection;
+
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import lombok.ToString;
+
+import javax.validation.constraints.Max;
+import javax.validation.constraints.Min;
 
 /**
- * Tests for this package.
+ * Container class for RPC related properties.
  *
  * @author tgianos
  * @since 4.0.0
  */
-@ParametersAreNonnullByDefault
-package com.netflix.genie.web.spring.autoconfigure.dtos;
-
-import javax.annotation.ParametersAreNonnullByDefault;
+@RequiredArgsConstructor
+@Getter
+@EqualsAndHashCode(doNotUseGetters = true)
+@ToString(doNotUseGetters = true)
+public class GenieWebRpcInfo {
+    @Min(value = 1, message = "The minimum value for the RPC port is 1")
+    @Max(value = 65_535, message = "The maximum value for the RPC port is 65,535")
+    private final int rpcPort;
+}
